refactor(Text examples): update names to Content
