+++ packages/react-core/src/components/Content/examples/ContentUnorderedList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TextList, TextListItem } from '@patternfly/react-core';
 
-export const TextUnorderedList: React.FunctionComponent = () => (
+export const ContentUnorderedList: React.FunctionComponent = () => (
   <TextList>
     <TextListItem>In fermentum leo eu lectus mollis, quis dictum mi aliquet.</TextListItem>
     <TextListItem>Morbi eu nulla lobortis, lobortis est in, fringilla felis.</TextListItem>
