+++ packages/react-core/src/components/Content/examples/ContentWrapper.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { TextContent, Text, TextVariants } from '@patternfly/react-core';
 
-export const TextContentWrapped: React.FunctionComponent = () => (
+export const ContentWrapper: React.FunctionComponent = () => (
   <TextContent>
     <Text component={TextVariants.p}>
       Text with a component of type "p" still renders the same when wrapped with a TextContent.
