+++ packages/react-core/src/components/Content/index.ts
@@ -1,4 +1,4 @@
+export * from './Content';
 export * from './TextContent';
-export * from './Text';
 export * from './TextList';
 export * from './TextListItem';
